Move zip_helper out of boinc-autodock-vina folder. Remove lefotvers in the linux.yml file.
