+++ .github/workflows/linux.yml
@@ -1,6 +1,6 @@
 # This file is part of BOINC.
 # https://boinc.berkeley.edu
-# Copyright (C) 2022 University of California
+# Copyright (C) 2023 University of California
 #
 # BOINC is free software; you can redistribute it and/or modify it
 # under the terms of the GNU Lesser General Public License
@@ -69,25 +69,6 @@ jobs:
         run: |
           sudo apt-get -qq update
           sudo apt-get install -y gcc-aarch64-linux-gnu g++-aarch64-linux-gnu binutils-aarch64-linux-gnu qemu-user-static
-
-      - name: Install boinc-autodock-vina dependencies
-        if: ${{ success() && (matrix.configuration == 'x64-linux-static' || matrix.configuration == 'coverage-report') && matrix.app == 'boinc-autodock-vina' }}
-        run: |
-          echo "Installing ADFR Suite"
-          wget https://ccsb.scripps.edu/adfr/download/1038/ -O $HOME/ADFRsuite_x86_64Linux_1.0.tar.gz --no-check-certificate
-          tar zxvf $HOME/ADFRsuite_x86_64Linux_1.0.tar.gz -C $HOME
-          dir=$(pwd)
-          cd $HOME/ADFRsuite_x86_64Linux_1.0/
-          yes | ./install.sh -d $HOME/ADFRsuite -c 0
-          cd $dir
-          echo "$HOME/ADFRsuite/bin" >> $GITHUB_PATH
-          echo "Installing OpenBabel and RDKit"
-          sudo apt-get install -y openbabel libopenbabel-dev swig python3-rdkit librdkit1 rdkit-data
-          sudo ln -s /usr/include/openbabel3/ /usr/local/include/openbabel3
-          echo "$HOME/.local/bin" >> $GITHUB_PATH
-          export PATH=~/.local/bin/:"$PATH"
-          echo "Installing Meeko and Scipy"
-          pip3 install -U numpy openbabel meeko rdkit-pypi scipy matplotlib ipython jupyter pandas sympy nose
 
       - name: Check if build is running from origin repo
         if: ${{ success() && env.AWS_ACCESS_KEY_ID != 0 && env.AWS_SECRET_ACCESS_KEY != 0 }}
